mybatis-flex 测试 需要找bean转换工具

+++ platform/platform-system-bom/platform-system/src/main/java/com/zclcs/platform/system/service/OauthClientDetailsService.java
@@ -48,7 +48,7 @@ public interface OauthClientDetailsService extends IService<OauthClientDetails> 
      * @param oauthClientDetailsVo oauthClientDetailsVo
      * @return OauthClientDetailsVo
      */
-    Integer countOauthClientDetails(OauthClientDetailsVo oauthClientDetailsVo);
+    Long countOauthClientDetails(OauthClientDetailsVo oauthClientDetailsVo);
 
     /**
      * 新增
