+++ test/test-test-bom/test-test/src/main/java/com/zclcs/test/test/service/CompanyService.java
@@ -1,10 +1,10 @@
 package com.zclcs.test.test.service;
 
-import com.baomidou.mybatisplus.extension.service.IService;
+import com.mybatisflex.core.service.IService;
 import com.zclcs.cloud.lib.core.base.BasePage;
 import com.zclcs.cloud.lib.core.base.BasePageAo;
-import com.zclcs.test.test.api.bean.entity.Company;
 import com.zclcs.test.test.api.bean.ao.CompanyAo;
+import com.zclcs.test.test.api.bean.entity.Company;
 import com.zclcs.test.test.api.bean.vo.CompanyVo;
 
 import java.util.List;
@@ -21,7 +21,7 @@ public interface CompanyService extends IService<Company> {
      * 查询（分页）
      *
      * @param basePageAo {@link BasePageAo}
-     * @param companyVo {@link CompanyVo}
+     * @param companyVo  {@link CompanyVo}
      * @return {@link CompanyVo}
      */
     BasePage<CompanyVo> findCompanyPage(BasePageAo basePageAo, CompanyVo companyVo);
@@ -48,7 +48,7 @@ public interface CompanyService extends IService<Company> {
      * @param companyVo {@link CompanyVo}
      * @return 统计值
      */
-    Integer countCompany(CompanyVo companyVo);
+    Long countCompany(CompanyVo companyVo);
 
     /**
      * 新增
@@ -56,7 +56,7 @@ public interface CompanyService extends IService<Company> {
      * @param companyAo {@link CompanyAo}
      * @return {@link Company}
      */
-     Company createCompany(CompanyAo companyAo);
+    Company createCompany(CompanyAo companyAo);
 
     /**
      * 修改
@@ -64,7 +64,7 @@ public interface CompanyService extends IService<Company> {
      * @param companyAo {@link CompanyAo}
      * @return {@link Company}
      */
-     Company updateCompany(CompanyAo companyAo);
+    Company updateCompany(CompanyAo companyAo);
 
     /**
      * 删除
